# docs/reference/openapi.yaml
...
openapi: 3.1.0
info:
  title: Narad HTTP API
  version: v1
  summary: Produce, consume and manage topics on a Narad cluster over HTTP.
  license:
    name: Apache-2.0
    identifier: Apache-2.0
  x-narad-lead: |
    Look up every Narad HTTP endpoint: its parameters, request body, the
    grant it needs, the status codes it returns and a real example.
  x-narad-lead-example:
    request: |
      curl -i -u "$AUTH" "$NARAD/v1/topics/orders?partition=1"
    response: |
      HTTP/1.1 200 OK
      Content-Length: 456
      Content-Type: application/json
      Date: Mon, 28 Sep 2026 19:31:43 GMT

      {
        "name": "orders",
        "id": "681a429c0b683b2d",
        "partitions": 3,
        "retention_ms": 172800000,
        "visibility_timeout_ms": 30000,
        "max_in_flight_per_partition": 1024,
        "max_acked_ahead_per_partition": 1024,
        "created_at": 1790623903,
        "owner": "admin",
        "role": "parent",
        "children": ["orders-audit"],
        "schema_version": 0,
        "partition_stats": [
          {
            "index": 1,
            "segments": 1,
            "oldest_offset": 0,
            "next_offset": 2,
            "high_watermark": 2,
            "size_bytes": 166,
            "oldest_segment_at": 1790623903,
            "owner_node": "narad-0"
          }
        ]
      }
  x-narad-placeholders: |
    - `$NARAD` is the base URL of any node or of the load balancer in front
...
      off and `-u "$AUTH"` can be left out.

...

...

...

...

...

...

...

...

...

...

...

...

...
servers:
...
security:
  - basicAuth: []
...
paths:
...
  /v1/topics/{topic}/produce:
    parameters:
      - $ref: '#/components/parameters/TopicPath'
    post:
      operationId: produce
      tags: [Messages]
      summary: Produce a message
      description: |
        Stores the request body as one message. A
        [`202`](status-codes.md#status-202) means the node that answered has
        written the message to its [ingress WAL](glossary.md#ingress-wal)
        and synced it to disk; it then moves the message to the partition's
        owner in the background. What a `202` promises is in the
        [delivery contract](../understand/delivery-contract.md#what-202-means).

        The body is stored byte for byte. Send JSON with
        `Content-Type: application/json` and anything else with
        `Content-Type: application/octet-stream`. On a topic with a schema,
        the body must be one JSON text that the current schema accepts
        ([Schema validation rules](schema-rules.md#validation)).
      x-narad-grant: |
        `produce` on the topic.
      parameters:
        - name: key
          in: query
          description: |
            Routes the message: messages with the same key go to the same
            partition while the partition count is unchanged and its owner
            is up. This is not an ordering guarantee. Without a key, messages
            are spread round-robin.
          schema:
            type: string
        - name: partition
          in: query
          description: Pins the message to this partition. Wins over `key`.
          schema:
            type: integer
            minimum: 0
      requestBody:
        required: true
        description: The message, 1 byte to 1 MiB.
        content:
          application/json:
            schema: {}
          application/octet-stream:
            schema:
              type: string
              format: binary
      responses:
        '202':
          description: Accepted and synced to disk. The body is empty.
        '400':
          description: Empty body, `partition` out of range, `key` or `partition` given twice, or the body fails the topic's schema (from v3.1.0, that includes a body nested deeper than 256 levels).
        '401':
          description: Missing or wrong credentials.
        '403':
          description: No `produce` grant on the topic.
        '404':
          description: The topic does not exist.
        '409':
          description: 'The topic is a delay child, which only its parent can feed. Unreleased: the topic is a remote child''s stub, whose messages live on its remote (`remote child "<name>" lives on remote <remote>; consume it there`).'
        '413':
          description: The body is over 1 MiB.
        '415':
          description: No accepted `Content-Type` and no `X-Narad-Client` header.
        '429':
          description: Too many produces in flight for this user on this node, only when the operator set a produce cap ([Configuration reference](configuration.md#http)).
        '500':
          description: The node could not write to its ingress WAL. It answers every produce this way until it restarts.
        '503':
          description: 'New in v3.1.0. Nothing was checked or stored; retry through another node. Either this node is being decommissioned and takes no new produce (with `Retry-After: 1`), or the topic has a schema and every schema validation slot on the node stayed busy for 5 seconds.'
      x-narad-examples:
        - request: |
            curl -i -u "$AUTH" -X POST \
              "$NARAD/v1/topics/orders/produce?key=customer-42" \
              -H "Content-Type: application/json" \
              -d '{"order_id": "ord_123", "amount": 1250}'
          response: |
            HTTP/1.1 202 Accepted
            Date: Mon, 28 Sep 2026 19:31:43 GMT
            Content-Length: 0
...
  /v1/topics/{topic}/consume:
    parameters:
      - $ref: '#/components/parameters/TopicPath'
    get:
      operationId: consume
      tags: [Messages]
      summary: Consume messages
      description: |
        Without `offset`, takes the next available message and gives the
        caller a [lease](glossary.md#lease) on it for the topic's
        `visibility_timeout_ms`. Settle it with [ack](#ack) before the lease
        runs out, or the message is delivered again. The request can reach
        any node; the node gathers a message from whichever partition owner
        has one.

        With `partition` and `offset`, reads the record at that offset
        without taking a lease: a [replay](../build/replay.md). The answer
        has no `receipt_handle` and nothing needs acking.
      x-narad-grant: |
        `consume` on the topic.
      parameters:
        - name: wait
          in: query
          description: |
            How long to wait for a message before answering `204`, as a Go
            duration such as `500ms` or `10s`. Without it the answer is
            immediate. Values above the server's maximum (10 s by default,
            [Configuration reference](configuration.md#http)) are cut to it,
            and the response then carries `X-Narad-Wait-Clamped` with the
            value used.
          schema:
            type: string
            default: 0s
        - name: partition
          in: query
          description: Take messages from this partition only. Required with `offset`.
          schema:
            type: integer
            minimum: 0
        - name: offset
          in: query
          description: |
            Replay the record at this offset of `partition`. An offset past
            the end of the partition answers `204`; one that aged out of
            retention answers `410`.
          schema:
            type: integer
            minimum: 0
        - name: max
          in: query
          x-narad-since: v3.1.0
          description: |
            Take up to this many messages in one answer,
            `{"messages": [...]}`, each with its own lease. The request does
            not wait to fill `max`. Requires an `X-Narad-Client` header
            ([Required headers](#required-headers)). It counts as `max`,
            clamped to the cap, against the per-user consume cap. Cannot be combined with `offset`. A v3.0.1
            node ignores it and answers with one message in the
            single-message shape.
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: 'One message, or `{"messages": [...]}` with `max`.'
          headers:
            X-Narad-Wait-Clamped:
              description: Present when `wait` was cut to the server's maximum; the value used.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Message'
        '204':
          description: No message arrived within `wait`, or `offset` is past the end of the partition.
        '400':
          description: A parameter is invalid, `partition` is out of range, `offset` came without `partition`, `max` came with `offset`, or `max` came without an `X-Narad-Client` header.
        '401':
          description: Missing or wrong credentials.
        '403':
          description: No `consume` grant on the topic.
        '404':
          description: The topic does not exist.
        '409':
          description: 'Unreleased. The topic is a remote child''s stub, whose messages live on its remote (`remote child "<name>" lives on remote <remote>; consume it there`).'
        '410':
          description: Replay only. The record at `offset` aged out of retention or cannot be read.
        '421':
          description: The partition moved to another node while the request was served. Retry.
        '429':
          description: Too many consumes in flight for this user on this node.
        '500':
          description: The node could not read the partition, for example after a disk error.
        '502':
          description: With `partition`, the node forwarded the request to the partition's owner and got no answer.
          x-narad-origin: forwarding
        '503':
          description: With `partition`, the partition's owner is down.
      x-narad-examples:
        - title: one message
          request: |
            curl -i -u "$AUTH" "$NARAD/v1/topics/orders/consume?wait=5s"
          response: |
            HTTP/1.1 200 OK
            Content-Length: 179
            Content-Type: application/json
            Date: Mon, 28 Sep 2026 19:31:43 GMT

            {"topic":"orders","partition":1,"offset":0,"key":"customer-42","payload":{"order_id": "ord_123", "amount": 1250},"timestamp":1790623903,"receipt_handle":"1:0:181499699661178901"}
        - title: a batch (v3.1.0)
          request: |
            curl -i -u "$AUTH" -H 'X-Narad-Client: curl' \
              "$NARAD/v1/topics/orders/consume?max=10&wait=5s"
          response: |
            HTTP/1.1 200 OK
            Content-Length: 326
            Content-Type: application/json
            Date: Mon, 28 Sep 2026 19:31:43 GMT

            {"messages":[{"topic":"orders","partition":1,"offset":1,"key":"customer-42","payload":{"order_id": "ord_124"},"timestamp":1790623903,"receipt_handle":"1:1:6130356697557286029"},{"topic":"orders","partition":0,"offset":0,"key":"customer-7","payload":"hello","timestamp":1790623903,"receipt_handle":"0:0:2681459915124828679"}]}
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
  parameters:
    TopicPath:
      name: topic
      in: path
      required: true
      description: Name of the topic.
      schema:
        type: string
...
  schemas:
...
    Message:
      type: object
      description: 'One message. A batch consume answers `{"messages": [...]}` with one of these per message.'
      properties:
        topic:
          type: string
          description: Topic name.
        partition:
          type: integer
          description: Partition the message is stored in.
        offset:
          type: integer
          description: Position in the partition.
        key:
          type: string
          description: The produce key. Absent for a message produced without one.
        key_encoding:
          type: string
          enum: [base64]
          x-narad-since: v3.1.0
          description: '`base64` when the key is not valid UTF-8 and `key` holds it in base64.'
        payload:
          description: |
            The message as it was produced. Valid JSON comes back as JSON,
            other UTF-8 text as a JSON string, and anything else as a base64
            string with `payload_encoding`.
        payload_encoding:
          type: string
          enum: [base64]
          description: '`base64` when `payload` holds binary data in base64.'
        timestamp:
          type: integer
          description: When the message was committed to its partition, Unix seconds.
        receipt_handle:
          type: string
          description: The lease to settle with [ack](#ack), `partition:offset:nonce`. Absent on a replay.
...