# internal/transport/httpserver/handlers/respond.go
package handlers

import (
	"bytes"
	"context"
	"encoding/json"
	"errors"
	"log/slog"
	"net/http"
	"strconv"
	"sync"

	"github.com/debanganthakuria/narad/internal/domain/topic"
	"github.com/debanganthakuria/narad/internal/errs"
	"github.com/debanganthakuria/narad/internal/persistence/metastore"
)

// StatusClientClosedRequest is the (non-standard, nginx-originated)
// status for a request whose client went away before the reply. It is
// a client-side outcome and deliberately not a 5xx.
const StatusClientClosedRequest = 499

// jsonAppender lets response types serialize themselves without going
// through encoding/json reflection on the hot path.
type jsonAppender interface {
	AppendJSON([]byte) []byte
}

var responseBufferPool = sync.Pool{
	New: func() any { return new(bytes.Buffer) },
}

// maxPooledResponseBuffer caps the capacity of a buffer returned to the
// pool. A consume that served a 1 MiB payload would otherwise park a
// 1 MiB buffer in the pool for every idle P, and every later small
// response would pin it; buffers that grew past this are left to the GC.
const maxPooledResponseBuffer = 256 << 10

// releaseResponseBuffer returns buf to the pool unless it grew past
// maxPooledResponseBuffer.
func releaseResponseBuffer(buf *bytes.Buffer) {
	if buf.Cap() > maxPooledResponseBuffer {
		return
	}
	responseBufferPool.Put(buf)
}

// headerJSON is the Content-Type value of every JSON response. It is
// assigned to the header map directly rather than through Header.Set,
// which allocates a fresh one-element slice per call; net/http only
// reads it, and Add or Set replace the slice rather than writing into
// it (its capacity is its length).
var headerJSON = []string{"application/json"}

// WriteJSON encodes v as the JSON response body with the given
// status. Nil values produce a header-only response.
func (s *Set) WriteJSON(w http.ResponseWriter, status int, v any) {
	w.Header()["Content-Type"] = headerJSON
	if v == nil {
		w.WriteHeader(status)
		return
	}

	buf := responseBufferPool.Get().(*bytes.Buffer)
	buf.Reset()
	defer releaseResponseBuffer(buf)

	if appender, ok := v.(jsonAppender); ok {
		buf.Write(appender.AppendJSON(buf.AvailableBuffer()))
		buf.WriteByte('\n')
	} else if err := json.NewEncoder(buf).Encode(v); err != nil {
		s.Deps.Logger.Error("encode response", "err", err)
		w.WriteHeader(status)
		return
	}

	w.Header().Set("Content-Length", strconv.Itoa(buf.Len()))
	w.WriteHeader(status)
	if _, err := w.Write(buf.Bytes()); err != nil {
		s.Deps.Logger.Error("write response", "err", err)
	}
}

// WriteMessage writes msg as the JSON response body with the given
// status: the bytes and headers WriteJSON(w, status, msg) writes, for
// the one response every delivered record gets. Passing the message by
// pointer keeps it off the heap; boxed into WriteJSON's interface it
// escaped on every local consume.
func (s *Set) WriteMessage(w http.ResponseWriter, status int, msg *topic.Message) {
	buf := responseBufferPool.Get().(*bytes.Buffer)
	buf.Reset()
	defer releaseResponseBuffer(buf)
	buf.Write(msg.AppendJSON(buf.AvailableBuffer()))
	buf.WriteByte('\n')

	h := w.Header()
	h["Content-Type"] = headerJSON
	h.Set("Content-Length", strconv.Itoa(buf.Len()))
	w.WriteHeader(status)
	if _, err := w.Write(buf.Bytes()); err != nil {
		s.Deps.Logger.Error("write response", "err", err)
	}
}

// WriteError writes a `{"error": msg}` body with the given status.
func (s *Set) WriteError(w http.ResponseWriter, status int, msg string) {
	s.logServerError(status, msg)
	s.writeError(w, status, msg)
}

func (s *Set) writeError(w http.ResponseWriter, status int, msg string) {
	body := make([]byte, 0, len(msg)+14)
	body = append(body, `{"error":`...)
	body = topic.AppendJSONQuoted(body, msg)
	body = append(body, "}\n"...)

	w.Header().Set("Content-Type", "application/json")
	w.Header().Set("Content-Length", strconv.Itoa(len(body)))
	w.WriteHeader(status)
	if _, err := w.Write(body); err != nil {
		s.Deps.Logger.Error("write error response", "err", err)
	}
}

// undecidedMarker is a response writer that records an answer written
// without a decision on the change (the topic handlers' audit writer).
type undecidedMarker interface {
	MarkUndecided()
}

// logServerError logs 5xx responses only: 4xx errors are the client's
// fault and would just be noise.
func (s *Set) logServerError(status int, msg string, attrs ...slog.Attr) {
	if status < http.StatusInternalServerError {
		return
	}
	logAttrs := append([]slog.Attr{
		slog.Int("status", status),
		slog.String("error", msg),
	}, attrs...)
	s.Deps.Logger.LogAttrs(context.Background(), slog.LevelError, "http server error", logAttrs...)
}

// WriteBrokerError maps a broker error to the right HTTP status. op
// is a short verb ("produce", "ack", …) used in log messages and
// the generic 5xx body.
//
// Receipt-handle errors are mapped to discrete codes so clients can
// distinguish "the handle cannot be decoded" (400) from "no active
// reservation matches it" (410): the lease ran out, the message was
// redelivered, or the handle came from another topic or names a
// partition this topic does not have (a handle carries no topic).
// Out-of-order ack rejected by ackedAhead-cap maps to 503: the head is
// genuinely stuck and the client should back off.
//
// A 503 whose change may still take effect (errs.ErrOutcomeUnknown: the
// leader lost its leadership after appending it) is first reported to a
// writer that records outcomes, so the topic audit logs it as unknown
// rather than failed.
func (s *Set) WriteBrokerError(w http.ResponseWriter, op string, err error) {
	if errors.Is(err, errs.ErrOutcomeUnknown) {
		if m, ok := w.(undecidedMarker); ok {
			m.MarkUndecided()
		}
	}
	switch {
	case errors.Is(err, errs.ErrForbidden):
		// The leader's ownership re-check refused the caller.
		s.WriteError(w, http.StatusForbidden, err.Error())
	case errors.Is(err, errs.ErrTopicNotFound):
		s.WriteError(w, http.StatusNotFound, "topic not found")
	case errors.Is(err, errs.ErrTopicAlreadyExists):
		// The bare sentinel keeps its fixed text; a wrapped one says
		// why (a name that differs from an existing one only in case).
		msg := "topic already exists"
		if err != errs.ErrTopicAlreadyExists {
			msg = err.Error()
		}
		s.WriteError(w, http.StatusConflict, msg)
	case errors.Is(err, errs.ErrHandleMalformed):
		s.WriteError(w, http.StatusBadRequest, err.Error())
	case errors.Is(err, errs.ErrHandleStale):
		s.WriteError(w, http.StatusGone, err.Error())
	case errors.Is(err, errs.ErrAckedAheadFull):
		s.WriteError(w, http.StatusServiceUnavailable, err.Error())
	case errors.Is(err, errs.ErrUnavailable):
		s.WriteError(w, http.StatusServiceUnavailable, err.Error())
	case errors.Is(err, errs.ErrInvalidArgument),
		errors.Is(err, errs.ErrPartitionRequired):
		s.WriteError(w, http.StatusBadRequest, err.Error())
	case errors.Is(err, errs.ErrNotPartitionOwner):
		s.WriteError(w, http.StatusMisdirectedRequest, err.Error())
	case errors.Is(err, errs.ErrFanoutRoleConflict),
		errors.Is(err, errs.ErrFanoutChildLimit),
		errors.Is(err, errs.ErrFanoutSchemaMismatch),
		errors.Is(err, errs.ErrFanoutSchemaManaged),
		errors.Is(err, errs.ErrDelayedChildProduce),
		// A delay the parent's retention cannot buffer is a conflict
		// between two records (child delay vs parent retention), not a
		// malformed request: the documented contract is 409 on attach,
		// create-as-child, and a retention shrink alike.
		errors.Is(err, errs.ErrFanoutDelayTooLong),
		// A schema update whose precondition failed or whose topic has
		// no room for another version is a conflict with the topic's
		// current state, not a malformed request.
		errors.Is(err, errs.ErrSchemaVersionConflict),
		errors.Is(err, errs.ErrSchemaHistoryFull),
		// The topic changed under the request (deleted and recreated,
		// or grown) after the leader checked it, twice in a row.
		errors.Is(err, errs.ErrTopicChanged),
		// Every remote-child refusal wraps this umbrella (a stub that
		// lives on a remote, a second link to one target, a retention
		// floor): conflicts with the stub or the parent, not bad input.
		errors.Is(err, errs.ErrRemoteChildConflict),
		errors.Is(err, errs.ErrAlreadyExists):
		s.WriteError(w, http.StatusConflict, err.Error())
	case errors.Is(err, errs.ErrNotFound):
		s.WriteError(w, http.StatusNotFound, err.Error())
	case errors.Is(err, metastore.ErrRootProtected):
		s.WriteError(w, http.StatusForbidden, "the root account is protected")
	case errors.Is(err, context.Canceled):
		// The request context is cancelled when the client goes away
		// (net/http cancels it on connection close, including a client
		// that half-closed after sending its request). Nobody may read
		// the reply, but it must not be recorded and logged as a server
		// failure: the server did nothing wrong.
		s.WriteError(w, StatusClientClosedRequest, "client closed request")
	default:
		status := http.StatusInternalServerError
		msg := op + " failed"
		s.logServerError(status, msg, slog.String("op", op), slog.Any("err", err))
		s.writeError(w, status, msg)
	}
}